pull out resource, resource item, location, and location item
