+++ scheduler/models/__init__.py
@@ -1,11 +1,12 @@
+from .resource import Resource
+from .resource_item import ResourceItem
+from .location_item import LocationItem
+from .location import Location
 from .remains import (
     Schedulable,
     ResourceItem,
-    Resource,
     ActItem,
     ActResource,
-    LocationItem,
-    Location,
     WorkerItem,
     Worker,
     EventItem,
